+++ src/api/article.js
@@ -1,4 +1,4 @@
-import api from "./axios";
+import api from "./api";
 
 export const getDatas = async () => {
   const res = await api.get("/home");
