--- src/components/pages/Layout/Header/Bookmark/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBookmark } from "@fortawesome/free-regular-svg-icons";
-import { StBookmark } from './styled';
-
-const Bookmark = () => {
-  return (
-    <StBookmark>
-        <FontAwesomeIcon icon={faBookmark} />
-    </StBookmark>
-  )
-}
-
-export default Bookmark
