--- src/components/pages/Layout/Header/UserAvatar/index.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { StUserAvatar } from "./styled"
-
-const UserAvatar = () => {
-  return (
-    <StUserAvatar>
-      <img src="/assets/images/img-avatar.png" alt="" />
-    </StUserAvatar>
-  )
-}
-
-export default UserAvatar
