--- src/components/pages/home/SectionTitle/styled.js
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-export const StSectionTitle = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 20px 0;
-  font-weight: 700;
-  margin-top: 40px;
-
-  h1 {
-    font-size: 20px;
-    letter-spacing: -0.3px;
-  }
-
-  bottom {
-    font-size: 16px;
-  }
-`;
