--- src/components/shared/NumberComma/styled.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StNumberComma = styled.span`
-  margin-top: 2px;
-  font-size: ${(props) => props.size};
-  line-height: ${(props) => props.lineheight};
-  font-weight: ${(props) => props.weight};
-  font-family: "Inter", sans-serif;
-`;
